Conflicts:
	CIMS_SA/nbproject/private/configs/Run_in_Browser.properties
	CIMS_SA/nbproject/private/private.properties

+++ CIMS_Server 2.0/src/Global/DAO/UserDAOImpl.java
@@ -28,7 +28,6 @@ public class UserDAOImpl implements UserDAO {
     public int count() {
         Query q = em.createNamedQuery("User.count", User.class);
         Integer amount = (Integer)q.getSingleResult();
-        return amount.intValue();
-    }
-    
+        return amount;
+    }    
 }
